minor refactor on styles
