+++ src/ui-spa/src/components/case-registration/suspect-aliases-summary/index.module.scss
@@ -1,32 +1,5 @@
 .caseSuspectAliasesSummaryPage {
-  & h1 {
-    margin-bottom: 0;
-  }
   .summaryListWrapper {
     max-width: 65%;
-  }
-
-  .inputWrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 1rem;
-    margin-top: 1rem;
-    :global(.govuk-form-group) {
-      margin-bottom: 1.25rem;
-      width: 100%;
-    }
-
-    & h2 {
-      font-size: 1.1875rem;
-      margin: 0;
-    }
-  }
-  .errorSummaryWrapper {
-    margin-top: 1rem;
-    max-width: 30rem;
-    &:focus {
-      outline: 0.2rem solid #ffdd00;
-    }
   }
 }
