+++ src/ui-spa/src/components/case-registration/suspect-asn/index.tsx
@@ -14,7 +14,7 @@ import {
   getNextSuspectJourneyRoute,
   getPreviousSuspectJourneyRoute,
 } from "../../../common/utils/getSuspectJourneyRoutes";
-import styles from "./index.module.scss";
+import styles from "../index.module.scss";
 
 const SuspectASNPage = () => {
   type ErrorText = {
@@ -129,7 +129,7 @@ const SuspectASNPage = () => {
   const { suspectASNText = "" } = suspects[suspectIndex] || {};
 
   return (
-    <div className={styles.caseDetailsPage}>
+    <div>
       <BackLink to={previousRoute}>Back</BackLink>
       {!!errorList.length && (
         <div
