+++ src/ui-spa/src/components/case-registration/suspect-gender/index.tsx
@@ -16,7 +16,7 @@ import {
   getNextSuspectJourneyRoute,
   getPreviousSuspectJourneyRoute,
 } from "../../../common/utils/getSuspectJourneyRoutes";
-import styles from "./index.module.scss";
+import styles from "../index.module.scss";
 
 const SuspectGenderPage = () => {
   type ErrorText = {
@@ -182,7 +182,7 @@ const SuspectGenderPage = () => {
   } = suspects[suspectIndex] || {};
 
   return (
-    <div className={styles.caseDetailsPage}>
+    <div>
       <BackLink to={previousRoute}>Back</BackLink>
       {!!errorList.length && (
         <div
