--- src/ui-spa/src/components/case-registration/suspect-religion/index.module.scss
@@ -1,23 +1,0 @@
-.caseDetailsPage {
-  .inputWrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 1rem;
-    :global(.govuk-form-group) {
-      margin-bottom: 1.25rem;
-      width: 100%;
-    }
-    & h2 {
-      font-size: 1.1875rem;
-      margin: 0;
-    }
-  }
-  .errorSummaryWrapper {
-    margin-top: 1rem;
-    max-width: 30rem;
-    &:focus {
-      outline: 0.2rem solid #ffdd00;
-    }
-  }
-}
