+++ src/ui-spa/src/components/case-registration/suspect-sdo/index.tsx
@@ -14,7 +14,7 @@ import {
   getNextSuspectJourneyRoute,
   getPreviousSuspectJourneyRoute,
 } from "../../../common/utils/getSuspectJourneyRoutes";
-import styles from "./index.module.scss";
+import styles from "../index.module.scss";
 
 const SuspectSDOPage = () => {
   type ErrorText = {
@@ -132,7 +132,7 @@ const SuspectSDOPage = () => {
   } = suspects[suspectIndex] || {};
 
   return (
-    <div className={styles.caseDetailsPage}>
+    <div>
       <BackLink to={previousRoute}>Back</BackLink>
       {!!errorList.length && (
         <div
